Fix snippet list names containing `,`
