+++ .changeset/hungry-trees-perform.md
@@ -1,0 +1,5 @@
+---
+'@lblod/ember-rdfa-editor-lblod-plugins': patch
+---
+
+Fix bug with snippet list names containing a `,` displaying as multiple lists
